+++ packages/typespec-python/tasks.py
@@ -57,9 +57,6 @@ EMITTER_OPTIONS = {
     "type/model/usage/main.tsp": {
         "package-name": "typetest-model-usage",
     },
-    "type/model/record/main.tsp": {
-        "package-name": "typetest-model-record",
-    },
     "type/model/visibility/main.tsp": {
         "package-name": "typetest-model-visibility",
     },
@@ -71,6 +68,9 @@ EMITTER_OPTIONS = {
     },
     "type/property/value-types/main.tsp": {
         "package-name": "typetest-property-valuetypes",
+    },
+    "type/property/additional-properties/main.tsp": {
+        "package-name": "typetest-property-additionalproperties",
     },
     "type/union/main.tsp": {
         "package-name": "typetest-union",