+++ packages/typespec-python/CHANGELOG.md
@@ -18,7 +18,7 @@
 
 **Bug Fixes**
 
-- Support model extends from `Record<>` #1925
+- Support scenario `model MyModel is/extends Record<>` #1925
 
 ## 2023-05-19 - 0.9.2